Moving AEAD into utils
